#26 Rename Merge
