+++ lib/combine.js
@@ -1,1 +1,1 @@
-"use strict";var merge=require("./merge"),mixins=require("./transform-mixins");module.exports=function(r,e){var i=merge(r);return mixins(i,e)};
+"use strict";var merge=require("./merge-classes"),mixins=require("./transform-mixins");module.exports=function(e,r){var i=merge(e);return mixins(i,r)};
