+++ src/merge-classes.js
