+++ test/merge.test.js
@@ -1,21 +1,21 @@
-'use strict';
+'use strict'
 
-const expect = require('chai').expect;
-const merge = require('../src/merge');
+const expect = require('chai').expect
+const merge = require('../src/merge-classes')
 
 describe('Merge', () => {
   it('return the same object if it is passed one', () => {
     const before = {
       foo: 'bar',
       baz: 'fin',
-    };
+    }
     const after = {
       foo: 'bar',
       baz: 'fin',
-    };
+    }
 
-    expect(merge(before)).to.eql(after);
-  });
+    expect(merge(before)).to.eql(after)
+  })
 
   return it('merge objects passed through array', () => {
     const before = [
@@ -30,16 +30,16 @@ describe('Merge', () => {
           boxShadow: '0 1px 4px rgba(0,0,0,.24)',
         },
       },
-    ];
+    ]
     const after = {
       card: {
         background: '#fff',
         margin: '0',
         boxShadow: '0 1px 4px rgba(0,0,0,.24)',
       },
-    };
+    }
 
-    expect(merge(before)).to.eql(after);
-  });
+    expect(merge(before)).to.eql(after)
+  })
 
-});
+})
